cleanup

+++ bootstrap-0.sh
@@ -333,11 +333,7 @@ fi
 if test "$recovery_install" != "true"; then
     unmount_efi /mnt
     for i in /mnt/boot /mnt/efi2 /mnt/efi; do
-        if test -d "$i"; then
-            rm -r "$i"
-        elif test -L "$i"; then
-            rm "$i"
-        fi
+        if test -d "$i"; then  rm -r "$i"; elif test -L "$i"; then rm "$i"; fi
     done
     exit 0
 fi
